requirements.txt
pandas
streamlit
transformers
groq
torchvision 